renamed ui-components to ui-kit
