+++ src/components/CommentSection.js
@@ -12,11 +12,11 @@ import {XS_MAX} from 'styles/screen-sizes'
 import {logException} from '../config'
 import {login, requireAuth} from 'lib/auth'
 //components
-import UserHeader from 'ui-components/UserHeader'
-import InputWithProfile from 'ui-components/InputWithProfile'
-import TextButton from 'ui-components/TextButton'
-import FaTrash from 'ui-components/icons/FaTrash'
-import WarningDialog from 'ui-components/WarningDialog'
+import UserHeader from 'ui-kit/UserHeader'
+import InputWithProfile from 'ui-kit/InputWithProfile'
+import TextButton from 'ui-kit/TextButton'
+import FaTrash from 'ui-kit/icons/FaTrash'
+import WarningDialog from 'ui-kit/WarningDialog'
 
 const commentAvatarSize = '35px'
 const childCommentAvatarSize = '25px'
