+++ src/components/Dialog.js
@@ -11,7 +11,7 @@ import '../styles/css/react-modal.css'
 
 import Media from 'react-media'
 import Modal from 'react-modal'
-import ExitIcon from 'ui-components/icons/ExitIcon'
+import ExitIcon from 'ui-kit/icons/ExitIcon'
 
 const ExitBox = styled.div`
   position: fixed;
