+++ src/ui-kit/WarningDialog.js
