+++ src/ui-kit/icons/ExitIcon.js
