+++ src/ui-kit/icons/FaTrash.js
